requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.0.0
django-celery-beat==2.5.0
python-decouple==3.8
requests==2.31.0
celery==5.3.0
redis==5.0.0
twilio==8.10.0
psycopg2-binary==2.9.6
pytz==2023.3
APScheduler==3.10.4
python-dotenv==1.0.0
gunicorn==21.2.0


